Enable whole album playing

+++ features/catalog/application/src/main/java/com/diskin/alon/sonix/catalog/application/interfaces/AlbumRepository.kt
@@ -9,5 +9,5 @@ interface AlbumRepository {
 
     fun getAll(sorting: AlbumSorting): Observable<AppResult<List<Album>>>
 
-    fun get(id: Int): Observable<AppResult<Album>>
+    fun get(id: Long): Observable<AppResult<Album>>
 }
