+++ features/catalog/application/src/main/java/com/diskin/alon/sonix/catalog/application/model/AlbumDto.kt
@@ -1,6 +1,6 @@
 package com.diskin.alon.sonix.catalog.application.model
 
-data class AlbumDto(val id: Int,
+data class AlbumDto(val id: Long,
                     val name: String,
                     val artist: String,
                     val tracks: Int,
