+++ features/catalog/application/src/main/java/com/diskin/alon/sonix/catalog/application/model/AudioTrackDto.kt
@@ -1,5 +1,5 @@
 package com.diskin.alon.sonix.catalog.application.model
 
-data class AudioTrackDto(val id: Int,
+data class AudioTrackDto(val id: Long,
                          val name: String,
                          val artist: String)
