+++ features/catalog/application/src/main/java/com/diskin/alon/sonix/catalog/application/model/PlayTracksRequest.kt
@@ -1,3 +1,3 @@
 package com.diskin.alon.sonix.catalog.application.model
 
-data class PlayTracksRequest(val startIndex: Int, val ids: List<Int>)
+data class PlayTracksRequest(val startIndex: Int, val ids: List<Long>)
