+++ features/catalog/domain/src/main/java/com/diskin/alon/sonix/catalog/domain/Album.kt
@@ -1,6 +1,6 @@
 package com.diskin.alon.sonix.catalog.domain
 
-data class Album(val id: Int,
+data class Album(val id: Long,
                  val name: String,
                  val artist: String,
                  val tracks: Int,
