+++ features/catalog/presentation/src/main/java/com/diskin/alon/sonix/catalog/presentation/model/UiAlbum.kt
@@ -2,7 +2,7 @@ package com.diskin.alon.sonix.catalog.presentation.model
 
 import android.net.Uri
 
-data class UiAlbum(val id: Int,
+data class UiAlbum(val id: Long,
                    val name: String,
                    val artist: String,
                    val art: Uri)
