+++ features/catalog/presentation/src/main/java/com/diskin/alon/sonix/catalog/presentation/model/UiAlbumTrack.kt
@@ -1,4 +1,4 @@
 package com.diskin.alon.sonix.catalog.presentation.model
 
-data class UiAlbumTrack(val id: Int,
+data class UiAlbumTrack(val id: Long,
                    val name: String)
