+++ features/catalog/presentation/src/test/java/com/diskin/alon/sonix/catalog/presentation/AudioTracksViewModelTest.kt
@@ -164,7 +164,7 @@ class AudioTracksViewModelTest {
     @Test
     fun deleteTrackFromModel_WhenTrackDeleted() {
         // Given
-        val trackId = 1
+        val trackId = 1L
 
         // When
         viewModel.deleteTrack(trackId)
@@ -176,7 +176,7 @@ class AudioTracksViewModelTest {
     @Test
     fun updateViewError_WhenModelTrackDeleteFail() {
         // Given
-        val trackId = 1
+        val trackId = 1L
         val modelError = mockk<AppError>()
 
         // When
@@ -191,7 +191,7 @@ class AudioTracksViewModelTest {
     fun playModelTracks_WhenRequested() {
         // Given
         val index = 1
-        val ids = mockk<List<Int>>()
+        val ids = mockk<List<Long>>()
 
         every { playTracksUseCase.execute(any()) } returns Single.just(AppResult.Success(Unit))
